skill_collector_gemini/requirements.txt
Flask==2.3.2
google-generativeai==0.3.2
pdfminer.six==20221105
python-dotenv==1.0.0